+++ module/RapportActivite/src/RapportActivite/Entity/Db/RapportActivite.php
@@ -326,11 +326,11 @@ class RapportActivite implements ResourceInterface, HistoriqueAwareInterface
         return $this;
     }
 
-    public function getCalendrierPrevionnelFinalisationEnabled(): bool
+    public function getCalendrierPrevionnelFinalisationEnabled(AnneeUniv $anneeUnivCourante): bool
     {
         return
             !$this->estFinContrat() &&
-            $this->getThese()->getAnneesUnivInscriptionCount() >= 3; // à partir de la 3ème année d'inscription
+            $this->getThese()->getAnneesUnivInscriptionCount($anneeUnivCourante) >= 3; // à partir de la 3ème année d'inscription
     }
 
     public function getCalendrierPrevionnelFinalisation(): ?string