+++ module/RapportActivite/src/RapportActivite/Form/RapportActiviteAnnuelFormFactory.php
@@ -2,6 +2,7 @@
 
 namespace RapportActivite\Form;
 
+use Application\Service\AnneeUniv\AnneeUnivService;
 use Doctrine\ORM\EntityManager;
 use Psr\Container\ContainerInterface;
 use RapportActivite\Hydrator\RapportActiviteHydrator;
@@ -20,6 +21,10 @@ class RapportActiviteAnnuelFormFactory
 
         $form = new RapportActiviteAnnuelForm('rapport-activite');
         $form->setHydrator($hydrator);
+
+        /** @var \Application\Service\AnneeUniv\AnneeUnivService $anneeUnivService */
+        $anneeUnivService = $container->get(AnneeUnivService::class);
+        $form->setAnneeUnivService($anneeUnivService);
 
         return $form;
     }